requirements.txt
groq
google-generativeai
cohere
azure-cognitiveservices-speech
python-dotenv
SpeechRecognition
pyaudio
requests
